Cover #1461 with regression tests
