+++ tests/Event/Sqs/SqsEventTest.php
@@ -2,10 +2,8 @@
 
 namespace Bref\Test\Event\Sqs;
 
-use Bref\Context\Context;
 use Bref\Event\InvalidLambdaEvent;
 use Bref\Event\Sqs\SqsEvent;
-use Bref\Test\Fixture\SqsFakeHandler;
 use PHPUnit\Framework\TestCase;
 
 class SqsEventTest extends TestCase
@@ -42,18 +40,5 @@ class SqsEventTest extends TestCase
         $this->expectException(InvalidLambdaEvent::class);
         $this->expectExceptionMessage('This handler expected to be invoked with a SQS event. Instead, the handler was invoked with invalid event data');
         new SqsEvent([]);
-    }
-
-    public function test partial failure()
-    {
-        $event = json_decode(file_get_contents(__DIR__ . '/handler.json'), true);
-        $result = (new SqsFakeHandler)->handle($event, Context::fake());
-
-        self::assertSame($result, [
-            'batchItemFailures' => [
-                ['itemIdentifier' => '2'],
-                ['itemIdentifier' => '4'],
-            ],
-        ]);
     }
 }
